requirements.txt
asgiref==3.5.0
boto3==1.23.2
botocore==1.26.2
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.49.0
django-classy-tags==3.0.1
django-compression-middleware==0.4.1
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
django-restart==0.2.8.7
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
newrelic==7.10.0.175
oauthlib==3.2.0
Pillow==9.0.1
psycopg==3.0.14
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
touch==2020.12.3
tzdata==2022.1
urllib3==1.26.8
values==2020.12.3
waitress==2.1.1
whitenoise==6.1.0
zstandard==0.17.0